rename Clock namespace to PointInTime
