+++ src/PointInTime/Millisecond.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace Innmind\TimeContinuum\Clock;
+namespace Innmind\TimeContinuum\PointInTime;
 
 /**
  * @psalm-immutable
