+++ tests/Earth/PointInTime/NowTest.php
@@ -9,13 +9,13 @@ use Innmind\TimeContinuum\{
     Timezone,
     Format,
     Earth\ElapsedPeriod,
-    Clock\Year as YearInterface,
-    Clock\Month,
-    Clock\Day,
-    Clock\Hour,
-    Clock\Minute,
-    Clock\Second,
-    Clock\Millisecond,
+    PointInTime\Year as YearInterface,
+    PointInTime\Month,
+    PointInTime\Day,
+    PointInTime\Hour,
+    PointInTime\Minute,
+    PointInTime\Second,
+    PointInTime\Millisecond,
     Earth\Timezone\UTC,
     Earth\Period\Year,
     Earth\Period\Millisecond as MillisecondPeriod,
