+++ tests/Earth/PointInTime/PointInTimeTest.php
@@ -9,13 +9,13 @@ use Innmind\TimeContinuum\{
     Timezone,
     Format,
     Earth\ElapsedPeriod,
-    Clock\Year,
-    Clock\Month,
-    Clock\Day as DayInterface,
-    Clock\Hour,
-    Clock\Minute,
-    Clock\Second,
-    Clock\Millisecond,
+    PointInTime\Year,
+    PointInTime\Month,
+    PointInTime\Day as DayInterface,
+    PointInTime\Hour,
+    PointInTime\Minute,
+    PointInTime\Second,
+    PointInTime\Millisecond,
     Earth\Timezone\UTC,
     Earth\Period\Composite,
     Earth\Period\Day,
