+++ tests/PointInTime/DayTest.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Clock;
+namespace Tests\Innmind\TimeContinuum\PointInTime;
 
 use Innmind\TimeContinuum\{
-    Clock\Day,
-    Clock\Month,
-    Clock\Year,
+    PointInTime\Day,
+    PointInTime\Month,
+    PointInTime\Year,
 };
 use PHPUnit\Framework\TestCase;
