+++ tests/PointInTime/MillisecondTest.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Clock;
+namespace Tests\Innmind\TimeContinuum\PointInTime;
 
-use Innmind\TimeContinuum\Clock\Millisecond;
+use Innmind\TimeContinuum\PointInTime\Millisecond;
 use PHPUnit\Framework\TestCase;
 
 class MillisecondTest extends TestCase
