+++ tests/PointInTime/MinuteTest.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Clock;
+namespace Tests\Innmind\TimeContinuum\PointInTime;
 
-use Innmind\TimeContinuum\Clock\Minute;
+use Innmind\TimeContinuum\PointInTime\Minute;
 use PHPUnit\Framework\TestCase;
 
 class MinuteTest extends TestCase
