+++ tests/PointInTime/SecondTest.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Clock;
+namespace Tests\Innmind\TimeContinuum\PointInTime;
 
-use Innmind\TimeContinuum\Clock\Second;
+use Innmind\TimeContinuum\PointInTime\Second;
 use PHPUnit\Framework\TestCase;
 
 class SecondTest extends TestCase
